// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const NotificationChannel = {
  email: 'email',
  sms: 'sms',
  push: 'push'
} as const

export type NotificationChannel = (typeof NotificationChannel)[keyof typeof NotificationChannel]


export const NotificationStatus = {
  queued: 'queued',
  processing: 'processing',
  sent: 'sent',
  failed: 'failed',
  duplicate: 'duplicate'
} as const

export type NotificationStatus = (typeof NotificationStatus)[keyof typeof NotificationStatus]


export const LegalBasis = {
  treatment: 'treatment',
  consent: 'consent',
  legal_obligation: 'legal_obligation',
  legitimate_interest: 'legitimate_interest'
} as const

export type LegalBasis = (typeof LegalBasis)[keyof typeof LegalBasis]